process explorer - remove old code (part 2)
